db-monitoring-project/auth-server/requirements.txt
flask
flask_sqlalchemy
pyjwt
